# fix: optimize Range parsing and formatting