@wraithgar 2024-07-12
We generally squash-merge external PRs unless there's a compelling reason to keep each commit.  linting will pass if *either* the PR title or every commit has a valid conventional prefix.

Don't worry about keeping the commit messages "valid" as we go through this PR, as we'll be using the PR title if and when this lands.