@wraithgar 2024-07-12
Thanks for the help in reviewing this @kurtextrem.  Yes, that comment was referring to `s*` which is a classic ReDOS vector.  Historically we've used the `.split` approach to fix that, because it's rarely in a hot path where performance matters as much as it does here.   Good to know that `String.replace` is faster.